requirements.txt
cx_freeze
mysql-connector-python
